requirements.txt
flask>=3.0.0
python-socketio>=5.0.0
kvsqlite